Struct mlrefcell::MLRefCell

pub struct MLRefCell<T: ?Sized> { /* fields omitted */ }

A RefCell which can be "locked" to prevent future mutable borrows

Implementations

impl<T> MLRefCell<T>

pub const fn new(value: T) -> MLRefCell<T>

Creates a new MLRefCell containing value

pub fn into_inner(self) -> T

Consumes the MLRefCell, returning the wrapped value

pub fn replace(&self, val: T) -> T

Replaces the wrapped value with a new one, returning the old value, without deinitializing either one. This function corresponds to std::mem::replace.

pub fn replace_with<F: FnOnce(&mut T) -> T>(&self, f: F) -> T

Replaces the wrapped value with a new one computed from f, returning the old value, without deinitializing either one.

pub fn swap(&self, other: &RefCell<T>)

Swaps the wrapped value of self with the wrapped value of other, without deinitializing either one. This function corresponds to std::mem:swap.

impl<T: ?Sized> MLRefCell<T>

pub fn lock(&self) -> Result<(), AlreadyLockedError>

Locks the MLRefCell, so that future mutable borrows will result in an error.

pub fn is_locked(&self) -> bool

Returns whether the MLRefCell is locked.

pub fn borrow(&self) -> Ref<'_, T>

Immutably borrows the wrapped value, panicking if the value is currently mutably borrowed.

pub fn try_borrow(&self) -> Result<Ref<'_, T>, BorrowError>

Immutably borrows the wrapped value, returning a BorrowError if the value is currently mutably borrowed.

pub fn borrow_mut(&self) -> RefMut<'_, T>

Mutably borrows the wrapped value, panicking if the value is either currently borrowed or if the MLRefCell is locked.

pub fn try_borrow_mut(&self) -> Result<RefMut<'_, T>, BorrowMutError>

Mutably borrows the wrapped value, returning an error if the value is either currently borrowed or if the MLRefCell is locked.

pub fn as_ptr(&self) -> *mut T

Returns a raw pointer to the underlying data in this cell.

pub fn get_mut(&mut self) -> &mut T

Returns a mutable reference to the underlying data. This call borrows self mutably, so there is no need for runtime checks (including lock checks). This method is usually not what you want, so use borrow_mut instead.

pub unsafe fn try_borrow_unguarded(&self) -> Result<&T, BorrowError>

Immutably borrows the wrapped value, returning an error if the value is currently mutably borrowed. Unlike RefCell::borrow, this method is unsafe because it does not return a Ref, thus leaving the borrow flag untouched. Mutably borrowing the RefCell while the reference returned by this method is alive is undefined behaviour.